Move isomorphismLuksVF2 to its own method
This isomorphism implementation is now in a separate method. This way it
can be called also for finding the actual isomorphism, and the equals()
method may now choose a bit easier between different implementations
based on the size and connectivity of the graphs, if more
implementations were to become available in the future.