Consistent use of 'me' and 'other'
Previously the code was written with the knowledge that numVerticesMe is
equal to numVerticesOther and numArcsMe to numArcsOther. This is
correct, but was inconsistently applied and makes the code harder to
debug if this assumption is broken (such as for subgraph isomorphism).
Therefore, the LuksVF2 algorithm now correctly uses numVerticesOther and
numArcsOther where it refers to the number of vertices and arcs in the
other graph.