Fix pseudo-recursion in Gabow's algorithm
The pseudo-recursion in Gabow's algorithm required a more sophisticated form of state tracking than was in the previous implementation, since it wasn't simple tail recursion. It isn't pretty, but now it also tracks a state in the form of one of two actions that need to be performed at each iteration. This indicates the pre-order action and the post-order action of the recursion.